backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
motor==3.6.0
pymongo==4.10.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.6
pypdf==5.1.0
python-docx==1.1.2
openai==1.59.5
firebase-admin==6.4.0
python-magic-bin==0.4.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3

# LangChain for RAG pipeline
langchain==0.3.13
langchain-groq==0.2.1
langchain-community==0.3.13
langchain-huggingface==0.1.2
sentence-transformers==3.3.1
faiss-cpu==1.9.0
tiktoken==0.8.0
